as_codec_dpkg

Convert a tibble to a data package (dpkg) object in R while checking it against CoDEC data specifications:

Usage

as_codec_dpkg(
  x,
  name,
  version,
  title = character(),
  description = character(),
  homepage = character()
)

Arguments

x

data.frame or tibble meeting CoDEC data specifications above

name

see dpkg::as_dpkg()

version

see dpkg::as_dpkg()

title

see dpkg::as_dpkg()

description

see dpkg::as_dpkg()

homepage

see dpkg::as_dpkg()

Value

a dpkg object

Details

The data must include a census tract identifier column (i.e., census_tract_id_2000, census_tract_id_2010, or census_tract_id_2020). The column must contain 11-digit GEOID identifiers for every census tract in Hamilton County, OH.

The data includes a year column (year), an integer year representing the vintage of the data (e.g. 2021). The data can optionally include a month column (month), an integer month of the year.

Data must be structured in a tidy format such that each row is an observation for a specific census tract at a specific year (and month).