Cincy neighborhood geographies

CAGIS data (see install_cagis_data()) provides community council boundaries, but these boundaries can overlap and do not align with census geographies or ZIP codes. By default, the statistical neighborhood approximations are instead returned, which are calculated by aggregating census tracts into 50 matching neighborhoods.

Usage

cincy_neighborhood_geo(
  geography = c("statistical_neighborhood_approximations", "community_council")
)

Arguments

geography

which type of cincy neighborhood geography to return

Value

a simple features object with a geographic identifier column (geoid) and a geometry column (s2_geography)

Examples

cincy_neighborhood_geo("statistical_neighborhood_approximations")
#> Simple feature collection with 50 features and 1 field
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -84.71216 ymin: 39.05206 xmax: -84.36954 ymax: 39.22101
#> Geodetic CRS:  WGS 84
#> # A tibble: 50 × 2
#>    geoid                                                            s2_geography
#>    <chr>                                                      <MULTIPOLYGON [°]>
#>  1 South Fairmount (((-84.57299 39.13404, -84.57322 39.13188, -84.57328 39.1313…
#>  2 Roselawn        (((-84.45447 39.18302, -84.45447 39.18307, -84.45462 39.1830…
#>  3 Mt. Lookout     (((-84.41474 39.13997, -84.41493 39.13998, -84.41501 39.1392…
#>  4 Westwood        (((-84.56772 39.15794, -84.56827 39.15795, -84.56888 39.1580…
#>  5 Mt. Adams       (((-84.49349 39.10796, -84.4936 39.10808, -84.49359 39.10824…
#>  6 Millvale        (((-84.54839 39.13769, -84.54846 39.13777, -84.54849 39.1378…
#>  7 North Fairmount (((-84.56189 39.14178, -84.56229 39.14177, -84.5628 39.1418,…
#>  8 Sayler Park     (((-84.68073 39.10714, -84.68074 39.10715, -84.68075 39.1071…
#>  9 Carthage        (((-84.47734 39.20546, -84.47914 39.20546, -84.47934 39.2049…
#> 10 Pendleton       (((-84.5002 39.11306, -84.50034 39.11308, -84.50048 39.1131,…
#> # ℹ 40 more rows
cincy_neighborhood_geo("community_council")
#> Simple feature collection with 75 features and 1 field
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -84.7124 ymin: 39.052 xmax: -84.36884 ymax: 39.2211
#> Geodetic CRS:  WGS 84
#> # A tibble: 75 × 2
#>    geoid                                                            s2_geography
#>    <chr>                                                      <MULTIPOLYGON [°]>
#>  1 Paddock Hills                (((-84.47807 39.17352, -84.47824 39.17355, -84.…
#>  2 Corryville - Heights         (((-84.51412 39.1385, -84.51438 39.13852, -84.5…
#>  3 South Cumminsville           (((-84.5442 39.15189, -84.54402 39.15211, -84.5…
#>  4 Avondale - North Avondale    (((-84.50068 39.15816, -84.50088 39.15825, -84.…
#>  5 Corryville                   (((-84.50976 39.14284, -84.50979 39.14288, -84.…
#>  6 Spring Grove Village         (((-84.53863 39.18173, -84.53862 39.18172, -84.…
#>  7 CUF - Mount Auburn           (((-84.51411 39.12557, -84.51411 39.12557, -84.…
#>  8 Columbia Tusculum - East End (((-84.44062 39.11659, -84.44049 39.11662, -84.…
#>  9 Mount Lookout                (((-84.42838 39.12023, -84.4293 39.11988, -84.4…
#> 10 Riverside - Sayler Park      (((-84.6767 39.10206, -84.67652 39.10201, -84.6…
#> # ℹ 65 more rows