Cincy ZIP Code Tabulation Areas

Read ZIP Code Tabulation Areas (ZCTAs) geographies from the online Census TIGER/Line files into R

Usage

cincy_zcta_geo(vintage = as.character(2024:2013))

Arguments

vintage

a character vector of a year corresponding to the vintage of TIGER/Line data

Value

a simple features object with a geographic identifier column (geoid) and a geometry column (s2_geography)

Examples

cincy_zcta_geo()
#> Simple feature collection with 55 features and 1 field
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -84.8203 ymin: 39.02153 xmax: -84.25364 ymax: 39.31774
#> Geodetic CRS:  WGS 84
#> # A tibble: 55 × 2
#>    geoid                                                            s2_geography
#>    <chr>                                                      <MULTIPOLYGON [°]>
#>  1 45204 (((-84.62871 39.07849, -84.62776 39.07812, -84.62734 39.07798, -84.626…
#>  2 45252 (((-84.66512 39.29226, -84.66489 39.29185, -84.66425 39.29124, -84.663…
#>  3 45227 (((-84.42289 39.16567, -84.42254 39.16555, -84.42218 39.16536, -84.421…
#>  4 45202 (((-84.52413 39.10229, -84.52396 39.1021, -84.5237 39.10184, -84.52358…
#>  5 45206 (((-84.50191 39.12703, -84.50183 39.12668, -84.50005 39.12656, -84.500…
#>  6 45223 (((-84.58036 39.17165, -84.57969 39.17092, -84.57948 39.17076, -84.579…
#>  7 45229 (((-84.50576 39.14221, -84.50322 39.14206, -84.50328 39.14161, -84.502…
#>  8 45251 (((-84.63292 39.23024, -84.6322 39.23005, -84.63113 39.22978, -84.6308…
#>  9 45174 (((-84.32404 39.15266, -84.3232 39.15207, -84.32293 39.1519, -84.32288…
#> 10 45243 (((-84.39592 39.17062, -84.39528 39.17056, -84.39406 39.17045, -84.393…
#> # ℹ 45 more rows
cincy_zcta_geo("2018")
#> Simple feature collection with 54 features and 1 field
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -84.8203 ymin: 39.02153 xmax: -84.25367 ymax: 39.31774
#> Geodetic CRS:  WGS 84
#> # A tibble: 54 × 2
#>    geoid                                                            s2_geography
#>    <chr>                                                      <MULTIPOLYGON [°]>
#>  1 45001 (((-84.72605 39.14286, -84.72228 39.14148, -84.72042 39.13968, -84.720…
#>  2 45002 (((-84.70997 39.13664, -84.70982 39.13642, -84.70964 39.13612, -84.709…
#>  3 45030 (((-84.82016 39.22722, -84.82015 39.22046, -84.82015 39.21871, -84.820…
#>  4 45033 (((-84.76682 39.17676, -84.76681 39.17633, -84.76681 39.17588, -84.765…
#>  5 45041 (((-84.70786 39.20645, -84.70719 39.20654, -84.70643 39.20668, -84.705…
#>  6 45051 (((-84.64903 39.09576, -84.64878 39.0957, -84.64848 39.0956, -84.64776…
#>  7 45052 (((-84.70438 39.17207, -84.70379 39.16867, -84.70137 39.16923, -84.701…
#>  8 45111 (((-84.29704 39.19442, -84.29624 39.19446, -84.29563 39.19445, -84.295…
#>  9 45241 (((-84.45096 39.24057, -84.45076 39.24055, -84.4505 39.24114, -84.4502…
#> 10 45242 (((-84.39668 39.2323, -84.39663 39.23205, -84.39655 39.23172, -84.3965…
#> # ℹ 44 more rows