Coerce CoDEC data package into a simple features object

The name of the census tract column in the CoDEC data package is used to add the appropriate cincy geography.

Usage

codec_as_sf(x)

Arguments

x

a CoDEC data package

Value

a simple features object with a geographic identifier column (geoid) and a geometry column (s2_geography) in addition to the columns in x

Details

Tract identifers do not change across decennial censuses, but the digital representation of their boundaries may be improved over time. Here, data packages using 2010 tract identifers use the TIGER/Line 2019 tract shapefiles and data packages using 2020 tract identifiers use the TIGER/Line 2020 tract shapefiles

Examples

codec_as_sf(get_codec_dpkg("property_code_enforcements-v0.2.0"))
#> Simple feature collection with 67488 features and 6 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -84.8203 ymin: 39.02153 xmax: -84.25624 ymax: 39.31206
#> Geodetic CRS:  WGS 84
#> # A tibble: 67,488 × 7
#>    geoid                            s2_geography  year month n_violations n_addr
#>    <chr>                      <MULTIPOLYGON [°]> <dbl> <dbl>        <int>  <int>
#>  1 39061021602 (((-84.54739 39.24984, -84.54736…  1999     9           NA     NA
#>  2 39061021602 (((-84.54739 39.24984, -84.54736…  1999    10            1   1111
#>  3 39061021602 (((-84.54739 39.24984, -84.54736…  1999    11           NA     NA
#>  4 39061021602 (((-84.54739 39.24984, -84.54736…  1999    12           NA     NA
#>  5 39061021602 (((-84.54739 39.24984, -84.54736…  2000     1           NA     NA
#>  6 39061021602 (((-84.54739 39.24984, -84.54736…  2000     2           NA     NA
#>  7 39061021602 (((-84.54739 39.24984, -84.54736…  2000     3           NA     NA
#>  8 39061021602 (((-84.54739 39.24984, -84.54736…  2000     4           NA     NA
#>  9 39061021602 (((-84.54739 39.24984, -84.54736…  2000     5           NA     NA
#> 10 39061021602 (((-84.54739 39.24984, -84.54736…  2000     6           NA     NA
#> # ℹ 67,478 more rows
#> # ℹ 1 more variable: violations_per_addr <dbl>