read CoDEC tabular data resource

This function is shorthand for reading a CoDEC tabular data resource using read_tdr_csv() from the installed R package.

Usage

codec_data(name, geography = cincy::tract_tigris_2010, geometry = FALSE)

Arguments

name

name of installed codec tabular data resource

geography

a cincy geography object; codec data will be returned at this geography using cincy::interpolate with block-level population weights

geometry

return the merged geography object alongside the output as a simple features object?

Value

a tibble (codec tabular data resource), or simple features object when geometry = TRUE

Examples

codec_data("hamilton_traffic")
#> reading data...
#> # A tibble: 222 × 16
#>    census_tract_id_2010 n_parcels n_parcels_near_traffic n_parcels_near_truck_…¹
#>    <chr>                    <int>                  <int>                   <int>
#>  1 39061024800               1324                   1163                       0
#>  2 39061024901                409                    153                       7
#>  3 39061025002               2791                   1632                    1404
#>  4 39061025104                724                    409                     359
#>  5 39061025102               2409                   1353                     300
#>  6 39061024902               2734                    750                       1
#>  7 39061025101               1654                    861                       0
#>  8 39061025001               2318                   1329                     440
#>  9 39061008601                639                    404                       0
#> 10 39061008800                710                    554                       0
#> # ℹ 212 more rows
#> # ℹ abbreviated name: ¹​n_parcels_near_truck_traffic
#> # ℹ 12 more variables: n_households <dbl>, n_households_near_traffic <dbl>,
#> #   n_households_near_truck_traffic <dbl>, median_parcel_traffic <dbl>,
#> #   median_household_traffic <dbl>, median_parcel_truck_traffic <dbl>,
#> #   median_household_truck_traffic <dbl>, frac_households_near_traffic <dbl>,
#> #   frac_parcels_near_traffic <dbl>, …
codec_data("hamilton_traffic", cincy::neigh_cchmc_2020)
#> reading data...
#> interpolating data to neighborhood_2020...
#> # A tibble: 81 × 16
#>    neighborhood_2020     median_parcel_traffic median_household_traffic
#>    <chr>                                 <dbl>                    <dbl>
#>  1 Amberley                                 0                        0 
#>  2 Anderson Twp.-Newtown              4076360.                 4082516.
#>  3 Avondale                          21079500.                29044446.
#>  4 Blue Ash                           5730866.                 5735005.
#>  5 Bond Hill                         14653209.                21747984.
#>  6 Camp Washington                  142134372                156719637 
#>  7 Carthage                          12436691                 14055610.
#>  8 Cheviot                            8205357.                 8476028.
#>  9 Clifton                           23392227.                29486212.
#> 10 Colerain Twp.                      5394207.                 5444374.
#> # ℹ 71 more rows
#> # ℹ 13 more variables: median_parcel_truck_traffic <dbl>,
#> #   median_household_truck_traffic <dbl>, frac_households_near_traffic <dbl>,
#> #   frac_parcels_near_traffic <dbl>, frac_households_near_truck_traffic <dbl>,
#> #   frac_parcels_near_truck_traffic <dbl>, year <dbl>, n_parcels <dbl>,
#> #   n_parcels_near_traffic <dbl>, n_parcels_near_truck_traffic <dbl>,
#> #   n_households <dbl>, n_households_near_traffic <dbl>, …
codec_data("hamilton_landcover", geography = cincy::zcta_tigris_2010, geometry = TRUE)
#> reading data...
#> interpolating data to zcta_2010...
#> Simple feature collection with 57 features and 6 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 953086.2 ymin: 1838813 xmax: 1010799 ymax: 1893846
#> Projected CRS: NAD83(NSRS2007) / Conus Albers
#> First 10 features:
#>    zcta_2010 pct_green_2019 pct_impervious_2019 pct_treecanopy_2016  evi_2018
#> 1      45212       31.73511            57.86038            6.653412 0.3173193
#> 2      45204       59.41175            34.85815           26.840917 0.4065009
#> 3      45233       90.68587            15.40738           38.599786 0.4991435
#> 4      45216       51.20896            45.71411           14.182532 0.3965691
#> 5      45232       66.93201            31.17200           28.380006 0.4911445
#> 6      45251       77.48776            28.91199           22.004516 0.4709363
#> 7      45248       82.29174            23.88763           27.630290 0.4575770
#> 8      45215       68.25781            34.82849           23.024239 0.4414922
#> 9      45237       56.38734            43.94948           12.492480 0.4148861
#> 10     45240       74.30367            31.02104           20.753888 0.4378616
#>    year                       geometry
#> 1  2019 MULTIPOLYGON (((985627.7 18...
#> 2  2019 MULTIPOLYGON (((977449.5 18...
#> 3  2019 MULTIPOLYGON (((964704.8 18...
#> 4  2019 MULTIPOLYGON (((983383.5 18...
#> 5  2019 MULTIPOLYGON (((980269.4 18...
#> 6  2019 MULTIPOLYGON (((970337.7 18...
#> 7  2019 MULTIPOLYGON (((971363.4 18...
#> 8  2019 MULTIPOLYGON (((983611.1 18...
#> 9  2019 MULTIPOLYGON (((984422.1 18...
#> 10 2019 MULTIPOLYGON (((976570 1863...