Install CAGIS GIS database

This installs the CAGIS Open Data GIS database (.gdb) into the data directory for the codec package. Once downloaded, it will be reused across R sessions on the same computer. The geodatabase contains many layers that are updated quarterly. (Historical geodatabases are not available here.)

Usage

install_cagis_data(
  cagis_data_url =
    "https://www.cagis.org/Opendata/Quarterly_GIS_Data/CAGISOpenDataQ4_2024.gdb.zip"
)

Arguments

cagis_data_url

the url to the CAGIS Open Data .gdb.zip file; this changes quarterly, so check for something more recent if the file cannot be found

See also

Examples

options(timeout = max(2500, getOption("timeout")), download.file.method = "libcurl")
install_cagis_data()
#> [1] "/home/runner/.local/share/R/codec/CAGISOpenDataQ4_2024.gdb"
sf::st_layers(install_cagis_data())$name
#> Warning: GDAL Error 1: Error occurred in filegdbtable.cpp at line 864
#>  [1] "Addresses"                                         
#>  [2] "Auditor_Book_and_Page"                             
#>  [3] "Auditor_Book_and_Page__ATTACH"                     
#>  [4] "Buildings_Footprints"                              
#>  [5] "Cincinnati_Benchmarks"                             
#>  [6] "Cincinnati_City_Boundary"                          
#>  [7] "Cincinnati_City_Corporation_Line"                  
#>  [8] "Cincinnati_Community_Council_Neighborhoods"        
#>  [9] "Cincinnati_Historic_Districts"                     
#> [10] "Cincinnati_Owned_Properties"                       
#> [11] "Cincinnati_Statistical_Neighborhood_Approximations"
#> [12] "Cincinnati_Tax_Increment_Finance_Districts"        
#> [13] "Cincinnati_Zoning"                                 
#> [14] "Condo_Attributes"                                  
#> [15] "Condo_Polygons"                                    
#> [16] "Countywide_Zoning"                                 
#> [17] "Driveways"                                         
#> [18] "Fences"                                            
#> [19] "FEMA_FIRM"                                         
#> [20] "FEMA_Floodway"                                     
#> [21] "Hamilton_County_Boundary"                          
#> [22] "Hamilton_County_Benchmarks"                        
#> [23] "Monuments"                                         
#> [24] "Municipalites_and_Townships"                       
#> [25] "Parcel_Polygons"                                   
#> [26] "Parcel_Merged_Consolidations"                      
#> [27] "Parks_and_Green_Spaces"                            
#> [28] "Right_of_Way"                                      
#> [29] "Rivers_and_Lakes"                                  
#> [30] "Road_Pavement"                                     
#> [31] "Section_Lines"                                     
#> [32] "Sidewalks"                                         
#> [33] "Stream_Network"                                    
#> [34] "Street_Centerlines"                                
#> [35] "Subdivisions"                                      
#> [36] "Voter_Polling_Locations"                           
#> [37] "Voter_Precincts"