Cincy address geographies

CAGIS data (see install_cagis_data()) provides a list of all addresses in Hamilton County. Addresses are filtered for the following criteria:

use only addresses that have STATUS of ASSIGNED, USING, or REGISTERED and are not orphaned (ORPHANFLG == "N")

omit addresses with ADDRTYPEs that are milemarkers (MM), parks (PAR), infrastructure projects (PRJ), cell towers (CTW), vacant or commercial lots (LOT), and other miscellaneous non-residential addresses (MIS, RR, TBA)

s2 cell is derived from LONGITUDE and LATITUDE fields in CAGIS address database

Usage

cincy_addr_geo(packaged = TRUE)

Arguments

packaged

logical; use the data included with the package instead of (down)loading from the source data?

Value

a simple features object with columns cagis_address, cagis_address_place, cagis_address_type, cagis_s2, cagis_parcel_id, cagis_is_condo, and a geometry column (s2_geography)

Examples

cincy_addr_geo()
#> Simple feature collection with 322732 features and 6 fields (with 1277 geometries empty)
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -84.82008 ymin: 39.02659 xmax: -84.23815 ymax: 39.3797
#> Geodetic CRS:  WGS 84 (CRS84)
#> # A tibble: 322,732 × 7
#>    cagis_address cagis_address_place cagis_address_type cagis_s2 cagis_parcel_id
#>    <chr>         <chr>               <chr>              <s2_cel> <chr>          
#>  1 9483 ZOLA CT… NA                  BLD                -6.1312… 056100270082   
#>  2 208 MOHAWK S… NA                  BLD                -6.7019… 009500020062   
#>  3 11800 READIN… U-HAUL - 11800 REA… BLD                -6.1778… 060800280003   
#>  4 12050 PRINCE… STAPLES SRPINGDALE  BLD                -6.1729… 059900100003   
#>  5 6610 PFEIFFE… NA                  INF                -6.1813… 060300140062   
#>  6 6501 PFEIFFE… NA                  INF                -6.1813… 061201400358   
#>  7 4244 ROUND B… NA                  INF                -6.6887… 050001700036   
#>  8 8210 MARKET … NA                  BLD                -6.1820… 060300080279   
#>  9 8212 MARKET … NA                  BLD                -6.1820… 060300080279   
#> 10 9857 MONTGOM… NA                  BLD                -6.1821… 060300080300   
#> # ℹ 322,722 more rows
#> # ℹ 2 more variables: cagis_is_condo <lgl>, s2_geography <POINT [°]>