read CoDEC tabular data resource

This function is shorthand for reading a CoDEC tabular data resource using read_tdr_csv() from the installed R package.

Usage

codec_data(name, geography = cincy::tract_tigris_2010)

Arguments

name

name of installed codec tabular data resource

geography

a cincy geography object to spatially interpolate data to

Value

a frictionless tabular data resource (fr_tdr) object

Examples

codec_data("hamilton_property_code_enforcement")
#> 
#> ── hamilton_property_code_enforcement 
#> # version: 0.1.3
#> # title: Hamilton County Property Code Enforcement
#> # homepage: <https://geomarker.io/hamilton_property_code_enforcement>
#> # description: Number of property code enforcements per household by census
#> tract
#> # A tibble: 222 × 3
#>    census_tract_id_2010 violations_per_household  year
#>    <chr>                                   <dbl> <dbl>
#>  1 39061000200                              1.14  2022
#>  2 39061000700                             20.8   2022
#>  3 39061000900                             81.5   2022
#>  4 39061001000                             32.0   2022
#>  5 39061001100                             36.9   2022
#>  6 39061001600                            218.    2022
#>  7 39061001700                            114.    2022
#>  8 39061001800                            114.    2022
#>  9 39061001900                             19.4   2022
#> 10 39061002000                              9.20  2022
#> # ℹ 212 more rows
codec_data("hamilton_property_code_enforcement", geography = cincy::neigh_cchmc_2010)
#> 
#> ── hamilton_property_code_enforcement 
#> # version: 0.1.3
#> # title: Hamilton County Property Code Enforcement
#> # homepage: <https://geomarker.io/hamilton_property_code_enforcement>
#> # description: Number of property code enforcements per household by census
#> tract
#> # A tibble: 81 × 3
#>    neighborhood_2010     violations_per_household  year
#>    <chr>                                    <dbl> <dbl>
#>  1 Amberley                                 0.731  2022
#>  2 Anderson Twp.-Newtown                    0.587  2022
#>  3 Avondale                                54.0    2022
#>  4 Blue Ash                                13.7    2022
#>  5 Bond Hill                               70.8    2022
#>  6 Camp Washington                        172.     2022
#>  7 Carthage                               222.     2022
#>  8 Cheviot                                  1.65   2022
#>  9 Clifton                                 12.0    2022
#> 10 Colerain Twp.                            0.262  2022
#> # ℹ 71 more rows