CoDEC latest annual

CoDEC data is harmonized as the most recently available annual (or annual average of monthly) values at the census tract 2020 geography. The year represents the year this table was assembled and mainly used as a placeholder to this object can be used with functions that take a codec_tbl object (see ?as_codec_tbl). See the description metadata for the actual latest years used for each CoDEC table.

Usage

codec_latest_annual

Format

An object of class codec_tbl (inherits from tbl_df, tbl, data.frame) with 226 rows and 66 columns.

Examples

tibble::glimpse(codec_latest_annual)
#> Rows: 226
#> Columns: 66
#> $ census_tract_id_2020                           <chr> "39061000200", "3906100…
#> $ prop_poverty                                   <dbl> 0.4365, 0.0403, 0.3206,…
#> $ prop_recieved_public_assistance_income         <dbl> 0.5053, 0.0477, 0.2413,…
#> $ prop_family_households_with_single_householder <dbl> 0.7378, 0.1429, 0.2696,…
#> $ prop_employment_among_civilian_workforce       <dbl> 0.9076, 0.9662, 0.8495,…
#> $ prop_housing_units_occupied_by_renters         <dbl> 0.9811, 0.9285, 0.5654,…
#> $ prop_median_rent_to_income_ratio_among_renters <dbl> 41.4, 20.8, 23.9, 21.7,…
#> $ prop_housing_units_vacant                      <dbl> 0.1810, 0.0899, 0.1809,…
#> $ prop_white_and_not_hispanic_or_latino          <dbl> 0.0785, 0.7003, 0.6062,…
#> $ prop_black_and_not_hispanic_or_latino          <dbl> 0.8995, 0.1273, 0.3232,…
#> $ prop_white_and_hispanic_or_latino              <dbl> 0.0000, 0.0208, 0.0000,…
#> $ prop_black_and_hispanic_or_latino              <dbl> 0.0000, 0.0000, 0.0000,…
#> $ prop_health_insurance                          <dbl> 0.9330, 0.9032, 0.9593,…
#> $ prop_rent_burdened                             <dbl> 0.5987, 0.2109, 0.3142,…
#> $ prop_housing_conditions                        <dbl> 0.6063, 0.2310, 0.2677,…
#> $ prop_built_prior_1980                          <dbl> 0.8293, 0.0852, 0.0452,…
#> $ prop_limited_english_speaking                  <dbl> 0.0000, 0.0102, 0.0000,…
#> $ prop_adults_hs_edu                             <dbl> 0.8176, 0.9237, 0.7745,…
#> $ median_home_value                              <dbl> NaN, 360000, 470500, 28…
#> $ n_households                                   <dbl> 475, 1762, 833, 1077, 5…
#> $ n_households_children                          <dbl> 92, 80, 62, 114, 93, 83…
#> $ n_housing_units                                <dbl> 580, 1936, 1017, 1249, …
#> $ n_persons_under_18                             <dbl> 194, 91, 141, 301, 238,…
#> $ n_crime_incidents_property_per_month           <dbl> 2.750000, 33.750000, 7.…
#> $ n_crime_incidents_violent_per_month            <dbl> 2.5833333, 9.5000000, 4…
#> $ n_crime_incidents_other_per_month              <dbl> 0.08333333, 0.33333333,…
#> $ n_shots_fired_per_month                        <dbl> 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ n_reported_shootings_per_month                 <dbl> 0.16666667, 0.50000000,…
#> $ drive_time_avg                                 <dbl> 18.0, 13.0, 12.9, 12.0,…
#> $ prcnt_area_within_1mi_epa_npl_site             <dbl> 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ prcnt_area_within_1mi_epa_tri_site             <dbl> 100.00, 100.00, 100.00,…
#> $ prcnt_area_within_1m_epa_tsd_site              <dbl> 0.00, 0.00, 0.00, 0.00,…
#> $ prcnt_area_within_1mi_epa_rmp_site             <dbl> 100.0000, 71.0400, 100.…
#> $ prcnt_area_within_1mi_coal_mine                <dbl> 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ prcnt_area_within_1mi_lead_mine                <dbl> 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ prcnt_area_within_1_mi_greenspace              <dbl> 100, 100, 100, 100, 100…
#> $ prcnt_homes_built_before_1980                  <dbl> 93.9400, 69.6700, 88.12…
#> $ walkability_index_epa                          <dbl> 15.00000, 18.00000, 18.…
#> $ prcnt_area_within_1mi_railroad                 <dbl> 100.00000, 100.00000, 6…
#> $ prcnt_area_within_1mi_high_volume_road         <dbl> 100.00, 100.00, 100.00,…
#> $ prcnt_area_within_1mi_airport                  <dbl> 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ prcnt_area_huc12_watershed                     <dbl> 70.81000, 40.49000, 70.…
#> $ pct_green_2019                                 <dbl> 8.00000, 1.00000, 3.000…
#> $ pct_impervious_2019                            <dbl> 71.00000, 86.00000, 80.…
#> $ pct_treecanopy_2016                            <dbl> 0.00000, 0.00000, 1.000…
#> $ enhanced_vegatation_index_2018                 <dbl> 0.2802000, 0.1703000, 0…
#> $ fraction_apartment                             <dbl> 0.67000000, 0.14000000,…
#> $ fraction_assisted_housing                      <dbl> 0.33, NaN, NaN, 0.01, 0…
#> $ market_total_value                             <dbl> 4132660.00, 299000.00, …
#> $ acreage                                        <dbl> 7.1750000, 0.0000000, 0…
#> $ year_built                                     <dbl> 1961.000, 1880.000, 190…
#> $ online_market_total_value                      <dbl> 4853640.0, 315000.0, 29…
#> $ fraction_condominium                           <dbl> NaN, 0.7800000, 0.67000…
#> $ fraction_single_family_dwelling                <dbl> NaN, 0.0700000, 0.15000…
#> $ fraction_two_to_three_family_dwelling          <dbl> NaN, 0.000000, 0.050000…
#> $ fraction_homestead                             <dbl> NaN, 0.02000000, 0.0100…
#> $ fraction_other                                 <dbl> NaN, NaN, 0.04, NaN, 0.…
#> $ n_total_rooms                                  <dbl> 0.000000, 4.000000, 4.0…
#> $ n_bedrooms                                     <dbl> 0.000000, 1.000000, 1.5…
#> $ n_full_bathrooms                               <dbl> 0, 1, 1, 1, 2, 1, 2, 2,…
#> $ n_half_bathrooms                               <dbl> 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ n_property_code_enforcements                   <dbl> 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ aadtm_trucks_buses                             <dbl> 14950526, 1016202, 0, 0…
#> $ aadtm_tractor_trailer                          <dbl> 26837985, 1227620, 0, 0…
#> $ aadtm_passenger                                <dbl> 329708957, 20693798, 0,…
#> $ year                                           <int> 2025, 2025, 2025, 2025,…
glue::glue(attr(codec_latest_annual, "description"))
#> # CoDEC latest annual
#> CoDEC data is harmonized as the most recently available annual (or annual average of monthly) values at the census tract 2020 geography.
#> The year represents the year this table was assembled and mainly used as a placeholder so this object can be used with codec_*() functions.
#> The actual latest years used for each CoDEC table are:
#> acs_measures-latest_annual_2023
#> crime-latest_annual_2024
#> drivetime-latest_annual_2024
#> environmental_justice_index-latest_annual_2022
#> landcover-latest_annual_2019
#> parcel-latest_annual_2024
#> property_code_enforcements-latest_annual_2025
#> traffic-latest_annual_2020